fix(@embarkjs): fix unhandled promise rejection
Previously, after EmbarkJS checked for working storage dappConnections but couldn’t find any that worked, there was an unhandled promise rejection error being logged.

The cause of this was because there was a new error being thrown inside of the `async.detectSeries` callback function (the function that executes after `detectSeries` completes). This new error was being caught by a try/catch block inside of the `detectSeries` loop.

To fix the issue, there is no new error being thrown, instead the error is logged to the console.